Update Invoke-LenovoInventory.ps1
Moved chassis detection before battery condition in order to use it as condition for the battery condition. This is because the script exits if it doesn't find the WMI Lenovo_Battery class which makes it so all workstations fail to submit anything to Log Analytics as this class will never exist on workstations. At least it didn't for me on Windows 10 22H2 on a Lenovo P520 using Lenovo Vantage Commercial 10.2210.33.0. I did make sure to set the Lenovo policies to sync battery information on a daily basis and waited a couple of days in case it was just taking a long time to create the Lenovo_Battery class but it never came. So bypassing this requirement on workstations fixes it for me.